feat: 耗时展示优化 #9505
